refactor: settings page のモバイル概要 UI を feature 側へ切り出す
Instagram風アカウント概要のJSXをMobileAccountOverviewとして
features/settings/componentsへ移動し、page.tsxはPCリダイレクト制御と
合成のみに縮小(322行→約40行)。returnPath/settingsReturnQueryは
app層のsettings-return.tsに依存するため、featuresからのapp逆依存を
避けてpage.tsx側で計算しpropsで渡す。

routing.test.tsxはMobileAccountOverviewがbarrelを経由せず
'../constants'を直接importするため、SETTINGS_CATEGORIESのmock先を
constants.tsに変更しbarrelはimportOriginalで実体を維持する形に調整。

Refs #1532
